Refactoring, redesign of the interval_tree submodule (#11)
This pull request includes significant changes to the `biobit` project,
mainly focusing on the removal of deprecated code, updates to
dependencies, and improvements to the Python bindings. The most
important changes include the removal of deprecated modules and their
associated tests, updates to dependencies in `Cargo.toml` and
`pyproject.toml`, and modifications to the `interval_tree` submodule.

### Removal of Deprecated Code:
* Removed the entire `gindex` module.
* Removed the `repmasker` module.
* Removed the `IntoPyBundle` struct and its implementation from the
`bindings/bundle.rs` file.

### Dependency Updates:
* Updated the `noodles` dependency to version `0.96.0` in `Cargo.toml`.
* Removed `intervaltree` and `sortedcontainers` dependencies from
`pyproject.toml`.

### Changes to Python Bindings:
* Refactored the `Bits` and `BitsBuilder` classes in `interval_tree`
bindings to improve method signatures and functionality.
* Redesigned how interval tree results are stored.